Fixed SharpnessView not rendering in weapons page. Minor SharpnessView refactor.
