+++ app/src/main/res/layout/listitem_charge_blade_detail.xml
@@ -57,7 +57,7 @@
                     android:layout_marginStart="@dimen/margin_large"
                     tools:text="+0" />
 
-                <com.ghstudios.android.ui.general.SharpnessView
+                <com.gatheringhallstudios.mhworlddatabase.components.SharpnessView
                     android:id="@+id/sharpness_value"
                     android:layout_width="0dp"
                     android:layout_height="@dimen/row_height_small"
